Allow changing the path to the language server.
Some users might want to use a different Python language server binary
than the one automatically downloaded by the extension. For example, on
some distributions of Linux, the official binaries do not work, and a
patched version is needed. It can be installed from the repositories,
and this patch allows the user to force coc-python to use it.

This is accomplished by adding a setting `python.languageServerPath`,
which, if non-empty, is used as the language server binary path instead
of the one downloaded by the extension.